+++ src/render/overrides.rs
@@ -3,7 +3,7 @@ use windows::Win32::Graphics::Direct3D11::{
 };
 
 use super::{
-    scopes::{ScopeFrame, ScopeUnk3, ScopeUnk8, ScopeView},
+    scopes::{ScopeFrame, ScopeTransparent, ScopeTransparentAdvanced, ScopeUnk3, ScopeView},
     shader, DeviceContextSwapchain,
 };
 
@@ -108,6 +108,7 @@ impl ShaderOverrides {
 pub struct ScopeOverrides {
     pub view: ScopeView,
     pub frame: ScopeFrame,
+    pub transparent: ScopeTransparent,
     pub unk3: ScopeUnk3,
-    pub unk8: ScopeUnk8,
+    pub transparent_advanced: ScopeTransparentAdvanced,
 }